raise OSErrors from low-level dir functions

+++ rupypy/lib/dir.py
@@ -5,7 +5,7 @@ from rupypy.error import error_for_oserror
 from rupypy.module import ClassDef
 from rupypy.modules.enumerable import Enumerable
 from rupypy.objects.objectobject import W_Object
-from rupypy.utils.ll_dir import opendir, readdir, closedir
+from rupypy.utils.ll_dir import opendir, readdir, closedir, DIRP
 
 
 class W_Dir(W_Object):
@@ -13,22 +13,16 @@ class W_Dir(W_Object):
     classdef.include_module(Enumerable)
 
     def __del__(self):
-        if self.dirp:
+        if self.dirp != lltype.nullptr(DIRP):
             closedir(self.dirp)
 
     @classdef.method("initialize", path="str")
     def method_initialize(self, space, path):
-        msg = None
-        if not os.path.exists(path):
-            msg = "No such file or directory - %s" % path
-            w_errno = space.newint(errno.ENOENT)
-        elif not os.path.isdir(path):
-            msg = "Not a directory - %s" % path
-            w_errno = space.newint(errno.ENOTDIR)
-        if msg:
-            raise space.error(space.w_SystemCallError, msg, [w_errno])
         self.path = path
-        self.dirp = None
+        try:
+            self.dirp = opendir(path)
+        except OSError as e:
+            raise error_for_oserror(space, exc=e)
 
     @classdef.singleton_method("allocate")
     def method_allocate(self, space, args_w):
@@ -54,24 +48,21 @@ class W_Dir(W_Object):
 
     @classdef.singleton_method("delete", path="path")
     def method_delete(self, space, path):
-        assert path
+        if not path:
+            raise error_for_oserror(space, errno=errno.ENOENT)
         try:
             os.rmdir(path)
         except OSError as e:
-            raise error_for_oserror(space, e)
+            raise error_for_oserror(space, exc=e)
         return space.newint(0)
 
     @classdef.method("read")
     def method_read(self, space, args_w):
-        if not self.dirp:
-            dirp, errno = opendir(self.path)
-            if not dirp:
-                raise space.error(space.w_SystemCallError, "opendir failed", [space.newint(errno)])
-            self.dirp = dirp
-        filename, errno = readdir(self.dirp)
+        try:
+            filename = readdir(self.dirp)
+        except OSError as e:
+            raise error_for_oserror(space, exc=e)
         if not filename:
-            if errno == 0:
-                return space.w_nil
-            else:
-                raise space.error(space.w_SystemCallError, "readdir failed", [space.newint(errno)])
-        return space.newstr_fromstr(filename)
+            return space.w_nil
+        else:
+            return space.newstr_fromstr(filename)
